requirements.txt
python-dotenv==0.17.1',
        'python-dateutil==2.8.1',
        'python-dotenv==0.17.1',
        'pytz==2021.1',
        'requests==2.25.1',
        'requests-oauthlib==1.3.0',
        'requests-toolbelt==0.9.1',
        'rsa==4.7.2',
        's3transfer==0.3.6',
        'six==1.15.0',
        'SQLAlchemy==1.3.23',
        'SQLAlchemy-Utils==0.36.8',
        'urllib3==1.26.3',
        'Werkzeug==1.0.1',
        'wrapt==1.12.1',
        'zipp==3.4.0',
    ],
    extras_require={
        'dev': [
            'autopep8==1.5.6',
            'flake8==3.8.4',
            'isort==5.7.0',
            'mypy==0.812',
            'pytest==6.2.2',
            'pytest-cov==2.11.1',
            'pytest-mock==3.5.1',
            'pytest-sugar==0.9.4',
            'pytest-xdist==2.2.1',
            'tox==3.20.1',
        ],
    },
    entry_points={
        'console_scripts': [
            'app = app.cli:main',
        ],
    },
)